Add explicit cast and comment for altitude value

+++ src/pArduBridge/UAV_Model.cpp
@@ -1132,7 +1132,7 @@ bool create_missionPlan(std::vector<mavsdk::MissionRaw::MissionItem> &mission_pl
   mission_plan.push_back(make_mission_item_wp( // 1 takeoff
       lat_deg_home + 0.003429,                 // 382.4 meters north
       lon_deg_home - 0.000534,                 // 49 meters west
-      41,
+      static_cast<int32_t>(41),  // altitude_m expects integer (41.03 rounds to 41)
       15,
       MAV_FRAME_GLOBAL_RELATIVE_ALT,
       MAV_CMD_NAV_TAKEOFF));
